Fix the layout issue with noun detail card.
Realized that resizing the browser window fixed the layout issue. From
that, it was easy to see that forcing the layout to be redrawn at the
end of the #show action would do the trick. A bit of trial and error,
and I figured out that calling:

    doComponentLayout()

on the detail panel does the trick.